requirements.txt
# Core dependencies
aiohttp==3.9.5
requests==2.32.3
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Data processing
pandas==2.2.2
numpy==1.26.4

# Database
sqlalchemy==2.0.30
alembic==1.13.1
# psycopg2-binary==2.9.9  # Optional - only needed for PostgreSQL

# Web scraping
beautifulsoup4==4.12.3
lxml==5.2.2

# Optimization
pulp==2.8.0

# Machine learning
scikit-learn==1.5.0
xgboost==2.0.3

# Scheduling
schedule==1.2.1
apscheduler==3.10.4

# Logging and monitoring
loguru==0.7.2
prometheus-client==0.20.0

# Testing
pytest==8.2.1
pytest-asyncio==0.23.7
pytest-cov==5.0.0

# Development tools
black==24.4.2
flake8==7.0.0
mypy==1.10.0
pre-commit==3.7.1

# API documentation
pyyaml==6.0.1